refactor: split user and admin sections
