+++ release-notes.json
@@ -1,6 +1,6 @@
 {
   "version": "0.0.0",
-  "generated": "2025-08-29T01:00:00+00:00",
+  "generated": "2025-08-29T02:15:00+00:00",
   "changes": [
     {
       "id": "2025-08-28T09:35:00+00:00-docs-docs",
@@ -122,6 +122,17 @@
         "en": "Added login form variants page",
         "ru": "Добавлена страница вариантов форм входа"
       }
+    },
+    {
+      "id": "2025-08-29T02:15:00+00:00-docs-docs",
+      "timestamp": "2025-08-29T02:15:00+00:00",
+      "version": "0.0.0",
+      "type": "docs",
+      "scope": "docs",
+      "description": {
+        "en": "Documented user and admin code separation",
+        "ru": "Задокументировано разделение кода на пользователя и админа"
+      }
     }
   ],
   "daily": {
@@ -130,12 +141,12 @@
       "ru": "Страница release notes показывает только английский текст; уточнены правила и стиль; разделены меню пользователя и админа; добавлен сворачиваемый пользовательский сайдбар; затемнён сайдбар и упорядочены иконки; реализован тёмный режим"
     },
     "2025-08-29": {
-      "en": "Aligned page titles with headings; fixed release notes layout; added login form variants page",
-      "ru": "Синхронизированы тайтлы с заголовками; исправлено отображение release notes; добавлена страница вариантов форм входа"
+      "en": "Aligned page titles with headings; fixed release notes layout; added login form variants page; documented user/admin code separation",
+      "ru": "Синхронизированы тайтлы с заголовками; исправлено отображение release notes; добавлена страница вариантов форм входа; задокументировано разделение кода на пользователя и админа"
     }
   },
   "ultrashort": {
-    "en": "Release notes English-only; split user/admin menus; collapsible sidebar; dark mode; unified page titles; fixed release notes layout; login form variants page",
-    "ru": "Release notes только на английском; разделены меню; добавлен сворачиваемый сайдбар; тёмный режим; унификация тайтлов; исправлено отображение release notes; страница вариантов форм входа"
+    "en": "Release notes English-only; split user/admin menus; collapsible sidebar; dark mode; unified page titles; fixed release notes layout; login form variants page; documented code separation",
+    "ru": "Release notes только на английском; разделены меню; добавлен сворачиваемый сайдбар; тёмный режим; унификация тайтлов; исправлено отображение release notes; страница вариантов форм входа; задокументировано разделение кода"
   }
 }
