+++ src/user/app/layout.css
@@ -1,0 +1,9 @@
+.layout {
+  display: flex;
+  width: 100%;
+}
+
+.layout > main {
+  flex: 1;
+  padding: 16px;
+}
