+++ src/user/app/layout.jsx
@@ -1,0 +1,14 @@
+import { Outlet } from 'react-router-dom'
+import BarLeftUser from './barLeftUser.jsx'
+import './layout.css'
+
+export default function Layout() {
+  return (
+    <div className="layout">
+      <BarLeftUser />
+      <main>
+        <Outlet />
+      </main>
+    </div>
+  )
+}
